use MVM_platform_cpu_count to implement cpucores op, but use uv impl
